Sample of interface use case
Use sendMsgToAll() function to send a message to all clients.
modification of backender server is only required if you want to change the message being relayed. If you do not need any changes, the message will be sent as is by default (how chat msg works currently).
Similar on client side, default there is to log the header and body.
There is also a sendToServer() method if you don't want the message to be related (probably good for save feature)